+++ word2vec/data_reader.py
