+++ word2vec/model.py
